+++ src/point/icons.js
@@ -1,0 +1,3 @@
+import {Icons} from '../const';
+export const getIcon = (type) =>
+  `<i class="trip-icon">${Icons.get(type)}</i>`;
