+++ src/point/index.js
@@ -1,0 +1,9 @@
+import {getIcon} from './icons';
+import {getOffers} from './offers';
+import {getSchedule} from './schedule';
+
+export {
+  getIcon,
+  getOffers,
+  getSchedule,
+};
